+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Managers/Controller_UI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc78e2baf7e1547f9b52770193aa5149
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
