+++ Metalord_btin/MetaLord/Assets/_Test/PSC/Scripts/Player/Controller_Physics.cs
@@ -108,7 +108,7 @@ public class Controller_Physics : MonoBehaviour
 
     // 12.21 SSC : NPC 대화중(stopState) 사격, 재장전 방지 위해 CanReload => !stopState 추가
     // 12.21 SSC : 상점창 사격, 재장전 방지 위해 CanReload => !storeUI.activeSelf 추가
-    public bool CanReload => !playingReloadAnimation && !OnClimb && !stopState && !storeUI.activeSelf;
+    public bool CanReload => !playingReloadAnimation && !OnClimb && !stopState && controller_UI.IsAnyUISetActiveFalse();
     public bool OnMultipleState => multipleState;
     public bool OnGround => groundContactCount > 0;
     public bool OnSteep => steepContactCount > 0;
@@ -174,18 +174,21 @@ public class Controller_Physics : MonoBehaviour
 
     RaycastHit aimHit;
 
-    //231219 배경택
-    [Header("UI 상점, 도감, 환경설정")]
-    [SerializeField] private GameObject canvases;
-    private GameObject storeUI; // 상점 UI 오브젝트
-    private GameObject recordUI; // 도감 UI 오브젝트
-    private GameObject readyMenuUI; // 대기모드 UI 오브젝트
-    private GameObject explainUI; // 도움말 UI 오브젝트    
-    private GameObject firstCoinExPlainUI; // 첫 코인 도움말 UI 오브젝트
-    private GameObject savingUI; // 첫 코인 도움말 UI 오브젝트
-    private GameObject firstKeyExplainUI; // 첫 코인 도움말 UI 오브젝트
-    private bool canInput = true; // 입력 가능여부
-    private const float INPUT_DELAYTIME = 0.3f; // 입력 후 대기 시간
+    [Header("UI 컨트롤러")]
+    [SerializeField] private Controller_UI controller_UI;
+    
+    ////231219 배경택
+    //[Header("UI 상점, 도감, 환경설정")]
+    //[SerializeField] private GameObject canvases;
+    //private GameObject storeUI; // 상점 UI 오브젝트
+    //private GameObject recordUI; // 도감 UI 오브젝트
+    //private GameObject readyMenuUI; // 대기모드 UI 오브젝트
+    //private GameObject explainUI; // 도움말 UI 오브젝트    
+    //private GameObject firstCoinExPlainUI; // 첫 코인 도움말 UI 오브젝트
+    //private GameObject savingUI; // 첫 코인 도움말 UI 오브젝트
+    //private GameObject firstKeyExplainUI; // 첫 코인 도움말 UI 오브젝트
+    //private bool canInput = true; // 입력 가능여부
+    //private const float INPUT_DELAYTIME = 0.3f; // 입력 후 대기 시간
 
     #region Animator Hash
     private readonly int IdleTimeHash = Animator.StringToHash("IdleTime");
@@ -224,120 +227,120 @@ public class Controller_Physics : MonoBehaviour
         canFire = false;
         fireDelay = StartCoroutine(fireDelayRoutine(fireDelayTime));
 
-        // 231231 배경택
-        storeUI = Utility.FindChildObj(canvases, "StoreCanvas");
-        recordUI = Utility.FindChildObj(canvases, "RecordCanvas");
-        readyMenuUI = Utility.FindChildObj(canvases, "ReadyCanvas");
-        explainUI = Utility.FindChildObj(canvases, "ExplainCanvas");
-        firstCoinExPlainUI = Utility.FindChildObj(canvases, "CoinExplainCanvas");
-        firstKeyExplainUI = Utility.FindChildObj(canvases, "KeyExplainCanvas");
-        savingUI = Utility.FindChildObj(canvases, "SavingCanvas");
+        //// 231231 배경택
+        //storeUI = Utility.FindChildObj(canvases, "StoreCanvas");
+        //recordUI = Utility.FindChildObj(canvases, "RecordCanvas");
+        //readyMenuUI = Utility.FindChildObj(canvases, "ReadyCanvas");
+        //explainUI = Utility.FindChildObj(canvases, "ExplainCanvas");
+        //firstCoinExPlainUI = Utility.FindChildObj(canvases, "CoinExplainCanvas");
+        //firstKeyExplainUI = Utility.FindChildObj(canvases, "KeyExplainCanvas");
+        //savingUI = Utility.FindChildObj(canvases, "SavingCanvas");
     }
 
     void Update()
     {
-        #region 상점, 도감, 환경설정 키 누를경우 _ 231219 배경택
-        if (canInput && !PlayerInteractNpc.isTalking)
-        {
-            if (reader.StoreKey) // 상점 키 누를 경우 _231219 배경택
-            {
-                if (storeUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    storeUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐    
-                }
-                else
-                {
-                    SwitchCameraLock(true);
-                    storeUI.SetActive(true);
-                    recordUI.SetActive(false);
-                    readyMenuUI.SetActive(false);
-                }
+        //#region 상점, 도감, 환경설정 키 누를경우 _ 231219 배경택
+        //if (canInput && !PlayerInteractNpc.isTalking)
+        //{
+        //    if (reader.StoreKey) // 상점 키 누를 경우 _231219 배경택
+        //    {
+        //        if (storeUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            storeUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐    
+        //        }
+        //        else
+        //        {
+        //            SwitchCameraLock(true);
+        //            storeUI.SetActive(true);
+        //            recordUI.SetActive(false);
+        //            readyMenuUI.SetActive(false);
+        //        }
 
-                StartCoroutine(DelayInput());
-            }
+        //        StartCoroutine(DelayInput());
+        //    }
 
-            if (reader.RecordKey) // 도감 키 누를 경우 _231219 배경택
-            {
-                if (recordUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    recordUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐
-                }
+        //    if (reader.RecordKey) // 도감 키 누를 경우 _231219 배경택
+        //    {
+        //        if (recordUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            recordUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐
+        //        }
 
-                else
-                {
-                    SwitchCameraLock(true);
-                    recordUI.SetActive(true);
-                    storeUI.SetActive(false);
-                    readyMenuUI.SetActive(false);
-                }
+        //        else
+        //        {
+        //            SwitchCameraLock(true);
+        //            recordUI.SetActive(true);
+        //            storeUI.SetActive(false);
+        //            readyMenuUI.SetActive(false);
+        //        }
 
-                StartCoroutine(DelayInput());
+        //        StartCoroutine(DelayInput());
 
-            }
+        //    }
 
-            if (IsAnyUISetActiveFalse() && reader.ReadyMenuKey) //대기메뉴 키 누를 경우 _231231 배경택
-            {
-                if (readyMenuUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    readyMenuUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐
-                }
-                else
-                {
-                    SwitchCameraLock(true);
-                    readyMenuUI.SetActive(true);
-                    recordUI.SetActive(false);
-                    storeUI.SetActive(false);
-                }
+        //    if (IsAnyUISetActiveFalse() && reader.ReadyMenuKey) //대기메뉴 키 누를 경우 _231231 배경택
+        //    {
+        //        if (readyMenuUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            readyMenuUI.SetActive(false); // 중복 버튼을 누를경우 꺼짐
+        //        }
+        //        else
+        //        {
+        //            SwitchCameraLock(true);
+        //            readyMenuUI.SetActive(true);
+        //            recordUI.SetActive(false);
+        //            storeUI.SetActive(false);
+        //        }
 
-                StartCoroutine(DelayInput());
+        //        StartCoroutine(DelayInput());
 
-            }
+        //    }
 
-            if (Input.GetKeyDown(KeyCode.Escape)) // 그냥 ESC키 누를경우 꺼짐 (환경설정키가 ESC로 되어있음에 따라 환경설정키는 조건에서 제외)
-            {
-                if (storeUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    storeUI.SetActive(false);
-                }
+        //    if (Input.GetKeyDown(KeyCode.Escape)) // 그냥 ESC키 누를경우 꺼짐 (환경설정키가 ESC로 되어있음에 따라 환경설정키는 조건에서 제외)
+        //    {
+        //        if (storeUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            storeUI.SetActive(false);
+        //        }
 
-                if (recordUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    recordUI.SetActive(false);
-                }               
+        //        if (recordUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            recordUI.SetActive(false);
+        //        }               
 
-                if (explainUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    explainUI.SetActive(false);
-                }
+        //        if (explainUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            explainUI.SetActive(false);
+        //        }
 
-                if(firstCoinExPlainUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    firstCoinExPlainUI.SetActive(false);
-                }
+        //        if(firstCoinExPlainUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            firstCoinExPlainUI.SetActive(false);
+        //        }
 
-                if(savingUI.activeSelf == true && savingUI.GetComponent<SavingCanvas>().isSaved)
-                {                    
-                    savingUI.SetActive(false);
-                    readyMenuUI.SetActive(true);
-                }
+        //        if(savingUI.activeSelf == true && savingUI.GetComponent<SavingCanvas>().isSaved)
+        //        {                    
+        //            savingUI.SetActive(false);
+        //            readyMenuUI.SetActive(true);
+        //        }
 
-                if(firstKeyExplainUI.activeSelf == true)
-                {
-                    SwitchCameraLock(false);
-                    firstKeyExplainUI.SetActive(false);
-                }
+        //        if(firstKeyExplainUI.activeSelf == true)
+        //        {
+        //            SwitchCameraLock(false);
+        //            firstKeyExplainUI.SetActive(false);
+        //        }
 
-                StartCoroutine(DelayInput());
-            }
-        }
-        #endregion
+        //        StartCoroutine(DelayInput());
+        //    }
+        //}
+        //#endregion
 
         //대화나 메뉴에서 stop시킴
         if (stopState)
@@ -411,27 +414,27 @@ public class Controller_Physics : MonoBehaviour
 
     }
 
-    bool IsAnyUISetActiveFalse()
-    {
-        if (storeUI.activeSelf
-            || recordUI.activeSelf
-            || explainUI.activeSelf
-            || firstCoinExPlainUI.activeSelf
-            || savingUI.activeSelf
-            || firstKeyExplainUI.activeSelf
-            ) return false;
+    //bool IsAnyUISetActiveFalse()
+    //{
+    //    if (storeUI.activeSelf
+    //        || recordUI.activeSelf
+    //        || explainUI.activeSelf
+    //        || firstCoinExPlainUI.activeSelf
+    //        || savingUI.activeSelf
+    //        || firstKeyExplainUI.activeSelf
+    //        ) return false;
 
-        return true;
-    }
+    //    return true;
+    //}
 
    
-    // 입력 대기시간 코루틴
-    IEnumerator DelayInput()
-    {
-        canInput = false; // 입력 불가
-        yield return new WaitForSecondsRealtime(INPUT_DELAYTIME); // 대기시간
-        canInput = true; // 입력 가능
-    }
+    //// 입력 대기시간 코루틴
+    //IEnumerator DelayInput()
+    //{
+    //    canInput = false; // 입력 불가
+    //    yield return new WaitForSecondsRealtime(INPUT_DELAYTIME); // 대기시간
+    //    canInput = true; // 입력 가능
+    //}
 
     private void FixedUpdate()
     {
